deployment/backend/requirements.txt
waitress
flask
pandas
numpy
scikit-learn
opencv-python-headless
tensorflow-cpu
pillow